% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isorig-package.r
\docType{package}
\name{isorig}
\alias{isorig}
\alias{isorig-package}
\title{isorig.}
\description{
isorig.
}

\examples{
library(isOrigin)
d <- getOrigData()
library(sp)
data("naMap") # load North America SpatialPolygonsDataFrame data in isOrigin package
d5 = subOrigData(taxon = "Turdus merula", mask = naMap) # No isotope data found in mask you choose!
d6 = subOrigData(taxon = c("Danaus plexippus", "Setophaga ruticilla", "Turdus migratorius"), mask = naMap)
d7 = subOrigData(group = "human", mask = naMap)
library(raster)
data(d2h_world)
d = subOrigData(taxon = c("Homo sapiens"), mask = naMap)[,2:4]
r = calRaster(known = d, isoscape = d2h_world, mask = naMap)
id = "smile"
d2H = -80
un = data.frame(id,d2H)
asn = pdRaster(r$isoscape.rescale,unknown=un,mask=naMap)

}
